Revert "feat: decorators 설정 추가 GlobalStyle, Design, IconLoader"
This reverts commit b83cf337a4465206b0b328e8931090d0182a5031.
